pyproject.toml
[project]
name = "pyeloverblik"
version = "0.1.1"
description = "Python client library for the Eloverblik API used by the Eloverblik Plus integration"
requires-python = ">=3.12"
license = "MIT"
dependencies = [
    "aiohttp>=3.9.0",
]

[tool.setuptools.packages.find]
include = ["pyeloverblik*"]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-homeassistant-custom-component",
    "ruff",
]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = [
    "B",      # flake8-bugbear
    "E",      # pycodestyle errors
    "F",      # pyflakes
    "I",      # isort
    "SIM",    # flake8-simplify
    "UP",     # pyupgrade
    "W",      # pycodestyle warnings
]

[tool.ruff.lint.isort]
force-sort-within-sections = true
known-first-party = ["custom_components.eloverblik_plus"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
